Move ViewModel to viewmodel package
