+++ app/src/main/java/com/uxstate/presentation/fav_photos_screen/FavoritePhotosScreen.kt
@@ -2,6 +2,7 @@ package com.uxstate.presentation.fav_photos_screen
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.ramcosta.composedestinations.annotation.Destination
@@ -13,8 +14,12 @@ import com.uxstate.util.LocalSpacing
 @OptIn(ExperimentalMaterial3Api::class)
 @Destination
 @Composable
-fun AstroShareScreen(astroPhotoEntity: AstroPhoto, navigator: DestinationsNavigator) {
-
+fun FavoritePhotosScreen(viewModel) {
     val spacing = LocalSpacing.current
-    AstroShareComposable(picture = astroPhotoEntity, modifier = Modifier.padding(spacing.spaceExtraSmall))
+    Scaffold() { padding ->
+
+        AstroShareComposable(picture = astroPhotoEntity, modifier = Modifier.padding(padding))
+    }
+
+
 }
