+++ app/src/main/java/com/uxstate/presentation/overview_screen/OverviewScreen.kt
@@ -20,6 +20,7 @@ import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.uxstate.R
 import com.uxstate.presentation.destinations.AstroShareScreenDestination
 import com.uxstate.presentation.components.AstroPhotoComposable
+import com.uxstate.presentation.viewmodel.OverviewViewModel
 import com.uxstate.util.LocalSpacing
 
 @OptIn(ExperimentalMaterial3Api::class)
